Pull request #3432: [BLUETOOTH-15265] APPLICATIONS: Fix MISRA cast issue
Merge in CONNINT/bluetooth_public from bugfix/BLUETOOTH-15265-rel to release/25.03.00

Squashed commit of the following:

commit adefe6460c3fbb91ad6aaff0badb78f31852163e
Author: Silviu Petria <silviu.petria@nxp.com>
Date:   Wed Mar 12 13:54:01 2025 +0200

    [BLUETOOTH-15265] APPLICATIONS: Fix MISRA cast issue

+++ application/common/app_advertiser.c
@@ -194,7 +194,7 @@ bleResult_t BluetoothLEHost_ReencryptAdvertisingData
             result = Gap_EncryptAdvertisingData(pAdvData, mKey, mIv, pOutput);
             if (result == gBleSuccess_c)
             {
-                mEncryptedAdvScanStruct.length = advLength;
+                mEncryptedAdvScanStruct.length = (uint8_t)advLength;
                 mEncryptedAdvScanStruct.adType = gAdEncryptedAdvertisingData_c;
                 mEncryptedAdvScanStruct.aData = pOutput;
                 result = Gap_SetExtAdvertisingData(advHandle, &mEncryptedAdvertisingData, NULL);
